% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_summary_block.R
\name{cr_bar}
\alias{cr_bar}
\title{Create bar in original typst}
\usage{
cr_bar(
  data,
  x_col,
  y_col,
  height = "1em",
  spacing = "0.4em",
  bar_color = "red",
  text_color = "luma(40)"
)
}
\arguments{
\item{data}{data.frame}

\item{x_col}{Character, name of X axis column}

\item{y_col}{Character, name of Y axis column}

\item{height}{Height of the bar}

\item{spacing}{Spacing between the bars}

\item{bar_color}{Color of the bars}

\item{text_color}{Color of the text}
}
\value{
Typst code chunk in quarto document
}
\description{
Create bar in original typst
}
