% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_summary_block.R
\name{cr_bar_text}
\alias{cr_bar_text}
\title{Create typst code chunk}
\usage{
cr_bar_text(
  text,
  text_size = "20pt",
  text_color = "luma(50)",
  text_tracking = "1pt",
  margin = "2em",
  inset_x = "0.5em",
  inset_y = "0.3em",
  outset_x = "0.3em",
  border_size = "0.5em",
  border_color = "blue"
)
}
\arguments{
\item{text}{Output text.}

\item{text_size}{Text size (default "20pt").}

\item{text_color}{Text color (default "luma(50)").}

\item{text_tracking}{Letter spacing (default "1pt").}

\item{margin}{Vertical margin (default "2em").}

\item{inset_x}{Horizontal padding (default "0.5em").}

\item{inset_y}{Vertical padding (default "0.3em").}

\item{outset_x}{Left outset (default "0.3em").}

\item{border_size}{Border thickness (default "0.5em").}

\item{border_color}{Border color (default "blue").}
}
\value{
Typst code chunk
}
\description{
Create typst code chunk
}
