% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_big_value.R
\name{cr_big_value}
\alias{cr_big_value}
\title{Create big value typst code chunk}
\usage{
cr_big_value(
  text = "5 year survival",
  value = "34.5\%",
  text_size = "15pt",
  value_size = "40pt",
  height_bottom = "5em",
  height_top = "3em",
  width = "10em",
  color = "blue"
)
}
\arguments{
\item{text}{Description text.}

\item{value}{Big value.}

\item{text_size}{Size of the text.}

\item{value_size}{Size of the value.}

\item{height_bottom}{Height of the bottom box.}

\item{height_top}{Height of the top box.}

\item{width}{Width of the box.}

\item{color}{Color of the background.}
}
\value{
Typst code chunk
}
\description{
Create big value typst code chunk
}
