% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_figure_value.R
\name{cr_figure_value}
\alias{cr_figure_value}
\title{Create figure box value}
\usage{
cr_figure_value(name = "men", value = "45.2\%")
}
\arguments{
\item{name}{Name of the figure.}

\item{value}{Value}
}
\value{
Typst code chunk used in Quarto document
}
\description{
Create figure box value
}
