% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_summary_block.R
\name{cr_rect_text}
\alias{cr_rect_text}
\title{Text surrounded by rectangle}
\usage{
cr_rect_text(
  text = "text",
  fill = "blue",
  text_color = "white",
  text_size = "1em",
  width = "auto",
  height = "auto"
)
}
\arguments{
\item{text}{Text.}

\item{fill}{Background color the rectangle.}

\item{text_color}{Color of the text.}

\item{text_size}{Size of the text.}

\item{width}{Width of the rectangle.}

\item{height}{Height of the rectangle.}
}
\value{
Quarto typst code chunk which output text surrounded by rectangle.
}
\description{
Text surrounded by rectangle
}
\examples{
cr_rect_text(text = "Incidence")
}
