% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_summary_block.R
\name{cr_summary_block}
\alias{cr_summary_block}
\title{Create typst code chunk}
\usage{
cr_summary_block(
  data,
  type = "incidence",
  lang = "cn",
  title_size = "20pt",
  label_size = "10pt",
  value_size = "18pt",
  fill = "red"
)
}
\arguments{
\item{data}{Data}

\item{type}{Type of summary block}

\item{lang}{Language used, \code{cn} for Chinese, \code{en} for English.}

\item{title_size}{The font size of the title.}

\item{label_size}{The font size of the value label.}

\item{value_size}{The font size of the value text.}

\item{fill}{The color filled in the box.}
}
\value{
Typst summary block
}
\description{
Create typst code chunk
}
