% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir_tree.R
\name{create_dir_tree}
\alias{create_dir_tree}
\title{Generate an overview of a manuscript factory.}
\usage{
create_dir_tree(path = ".", recurse = TRUE, ...)
}
\arguments{
\item{path}{A path to print the tree from.}

\item{recurse}{If TRUE recurse fully, if a positive number the number of
levels to recurse.}

\item{...}{Arguments passed on to dir_ls.}
}
\value{
Invisibly returns a character of the files and directories within the
desired folder.
}
\description{
Generate an overview of a manuscript factory.
}
