% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoxy_transformers.R
\name{epoxy_transformers}
\alias{epoxy_transformers}
\title{Manuscript Epoxy Transformers List}
\format{
A named list of functions.
}
\usage{
epoxy_transformers()
}
\description{
A list of custom transformer functions that can be used with \code{epoxy::epoxy_transform_set()}.
This is provided for optional integration with the epoxy package.
}
