% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_pkgs.R
\name{ls_pkgs}
\alias{ls_pkgs}
\title{List all R packages used in R scripts.}
\usage{
ls_pkgs(path = "scripts", include_install = FALSE)
}
\arguments{
\item{path}{Path of the R files or folders contains R files.}

\item{include_install}{Logical value, whether list installed packages or not.}
}
\value{
A character vector contains R package names.
}
\description{
List all R packages used in R scripts.
}
