% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_reports.R
\name{ls_reports}
\alias{ls_reports}
\title{List R Markdown and Quarto Reports}
\usage{
ls_reports(path = ".", manuscript_sources = "qmd_sources", pattern = NULL, ...)
}
\arguments{
\item{path}{The root directory where the search starts. Defaults to the current directory.}

\item{manuscript_sources}{The subdirectory within \code{path} where reports are located.}

\item{pattern}{An optional regex pattern to filter the listed reports.}

\item{...}{Additional arguments passed to \code{grep()}.}
}
\value{
A character vector of file paths to R Markdown and Quarto reports.
}
\description{
List R Markdown and Quarto Reports
}
