\name{manuscript-package}
\alias{manuscript-package}
\alias{manuscript}
\docType{package}
\title{
\packageTitle{manuscript}
}
\description{
\packageDescription{manuscript}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{manuscript}
\packageIndices{manuscript}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{manuscript}

Maintainer: \packageMaintainer{manuscript}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
