% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_cmc.R
\name{paged_plain}
\alias{paged_plain}
\title{Function for plain template}
\usage{
paged_plain(other_css = NULL, ...)
}
\arguments{
\item{other_css}{Add an other CSS}

\item{...}{Arguments passed to pagedown::html_paged#'}
}
\value{
A pagedown template
}
\description{
Function for plain template
}
