% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raw.R
\name{path_raw}
\alias{path_raw}
\alias{path_clean}
\alias{path_report}
\alias{path_figure}
\title{Path of the raw or clean data.}
\usage{
path_raw(filename)

path_clean(filename)

path_report(filename)

path_figure(filename)
}
\arguments{
\item{filename}{File name}
}
\value{
Character string contain the path of data.
}
\description{
Path of the raw or clean data.
}
