% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output.R
\name{save_output}
\alias{save_output}
\title{Save result to data_output folder.}
\usage{
save_output(x, name = "result")
}
\arguments{
\item{x}{Object}

\item{name}{File name.}
}
\value{
Created file.
}
\description{
Save result to data_output folder.
}
