% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_manuscript.R
\name{use_manuscript}
\alias{use_manuscript}
\title{Create a non-package project, i.e. a manuscript project}
\usage{
use_manuscript(
  name = "manuscript",
  path = "/Users/qc0824/Documents/1_Projects/",
  prefix = NULL,
  lang = "en",
  use_rproj = TRUE,
  use_gitignore = TRUE,
  use_readme = FALSE,
  type = "manuscript"
)
}
\arguments{
\item{name}{Manuscript project name.}

\item{path}{A path. If it exists, it is used. If it does not exist,}

\item{prefix}{Prefix of the manuscript name.}

\item{lang}{Language used for the manuscript project.}

\item{use_rproj}{Logical value, whether initiated with Rstudio project.}

\item{use_gitignore}{Logical value, whether add the gitignore file.}

\item{use_readme}{Logical value, whether add the readme md file.}

\item{type}{Type of the manuscript.}
}
\value{
Path to the newly created project or package, invisibly.
}
\description{
Create a non-package project, i.e. a manuscript project
}
