% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_qmd.R
\name{use_qmd}
\alias{use_qmd}
\title{Create a new Quarto document from a template}
\usage{
use_qmd(
  name = "a new quarto document",
  dir = "qsight",
  sub_dir = NULL,
  use_folder = TRUE,
  qmd_type = "blog",
  prefix = "date",
  include_yaml = TRUE,
  title = name,
  subtitle = "subtitle",
  lang = "cn",
  author = "Qiong Chen",
  date = Sys.Date(),
  overwrite = FALSE
)
}
\arguments{
\item{name}{Name of the Quarto document (default: "a new quarto document").}

\item{dir}{The directory of the document.}

\item{sub_dir}{The sub directory in the directory.}

\item{use_folder}{Create document in a dedicated folder (default: TRUE).}

\item{qmd_type}{Type of Quarto document (blog, slides, report, article; default: "blog").}

\item{prefix}{Naming prefix (NULL, "date", "rank", or custom string; default: "date").}

\item{include_yaml}{Include YAML header (default: TRUE).}

\item{title}{Title in the yaml header.}

\item{subtitle}{Subtitle in the yaml header.}

\item{lang}{Document language (en, cn, zh-cn; default: "cn").}

\item{author}{Document author (default: "Qiong Chen").}

\item{date}{Document date (default: current date).}

\item{overwrite}{Overwrite existing file (default: FALSE).}
}
\value{
Message indicating document creation path.
}
\description{
This function creates a new Quarto document with specified parameters,
including YAML headers and content based on the document type. It handles
directory creation, file naming, and existing file checks.
}
\examples{
\dontrun{
use_qmd(name = "My Report", qmd_type = "report")
}
}
