% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_script.R
\name{use_script}
\alias{use_script}
\title{Create R script in the manuscript.}
\usage{
use_script(script_type, name, prefix = "rank")
}
\arguments{
\item{script_type}{Part of the R script.}

\item{name}{Nmae of the R script.}

\item{prefix}{Prefix of the file name.}
}
\value{
Target path, invisibly.
}
\description{
Create R script in the manuscript.
}
