% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_source.R
\name{use_source}
\alias{use_source}
\title{Source all R files in a directory}
\usage{
use_source(path = "scripts", pattern = "\\\\.R$", quiet = TRUE)
}
\arguments{
\item{path}{The folder path containing the R files. Default is "scripts".}

\item{pattern}{The pattern used to identify R files. Default is "\\.R$" to match all .R files.}

\item{quiet}{Logical value. If TRUE, suppress messages. Default is FALSE.}
}
\value{
A vector of the file paths that were sourced.
}
\description{
Source all R files in a directory
}
