% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_zip2.R
\name{use_zip2}
\alias{use_zip2}
\title{Compress files}
\usage{
use_zip2(paths, zipfile, move_to = "outputs")
}
\arguments{
\item{paths}{Paths of the files to be compressed}

\item{zipfile}{Name of the output file}

\item{move_to}{Location of the compressed file}
}
\value{
Compressed file
}
\description{
Compress files
}
