% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript-package.R
\docType{package}
\name{manuscript-package}
\alias{manuscript}
\alias{manuscript-package}
\title{manuscript: Tools to Assist Statistical Report Writing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools to assist statistical report writing through the automated generation and management of analysis and reporting projects using RStudio Project workflows. The package enables the creation of standardized directory structures, Quarto (`.qmd`) documents, and README files, supporting reproducible and efficient report preparation.
}
\author{
\strong{Maintainer}: Qiong Chen \email{chenq08@126.com} [contributor, copyright holder]

}
\keyword{internal}
